requirements.txt
streamlit 
langchain_community
langchain_groq
PyPDF2
langchain_google_genai
langchain_core